+++ app/javascript/engine/control/movement.test.ts
@@ -421,10 +421,12 @@ describe("movement", () => {
     expect(game.moveState.addActions.length).toBe(0)
     game.moveState.move(3, 3)
     expect(showLaySmoke(game)).toBe(true)
-    expect(game.moveState.smoke).toBe(false)
+    expect(game.moveState.smoke).toBe(true)
     expect(game.moveState.addActions.length).toBe(1)
     expect(map.units[3][3].filter(u => u.ghost).length).toBe(1)
     expect(game.moveState.path.length).toBe(1)
+
+    game.moveState.smokeToggle()
 
     expect(game.gameState?.openHex(3, 2)).toBe(1)
     expect(game.gameState?.openHex(4, 3)).toBe(1)
@@ -3284,10 +3286,12 @@ describe("movement", () => {
     expect(game.moveState.addActions.length).toBe(0)
     game.moveState.move(3, 3)
     expect(showLaySmoke(game)).toBe(true)
-    expect(game.moveState.smoke).toBe(false)
+    expect(game.moveState.smoke).toBe(true)
     expect(game.moveState.addActions.length).toBe(1)
     expect(map.units[3][3].filter(u => u.ghost).length).toBe(1)
     expect(game.moveState.path.length).toBe(1)
+
+    game.moveState.smokeToggle()
 
     expect(game.gameState?.openHex(3, 2)).toBe(1)
     expect(game.gameState?.openHex(4, 3)).toBe(1)