+++ db/migrate/20260721173044_clear_games_six.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class ClearGamesSix < ActiveRecord::Migration[7.0]
+  def change
+    Game.delete_all
+  end
+end