+++ app/javascript/engine/control/state/BreakdownState.ts
@@ -12,7 +12,7 @@ export function breakdownCheck(game: Game): boolean {
   if (action.data.origin && action.data.origin.length > 0) {
     const id = action.data.origin[0].id
     const counter = game.findCounterById(id) as Counter
-    if (!counter.unit.breakdownRoll) { return false }
+    if (!counter || !counter.unit.breakdownRoll) { return false }
     if ("move" == action.data.action) { return true }
     if ("assault_move" === action.data.action) {
       if (action.data.add_action && action.data.add_action.length > 0 &&