+++ changelog.md
@@ -1,6 +1,11 @@
 # 0.115
 
 - Added tank crews, and special tank gun crew rules/abilities
+  - Cleared all games: tank crew make significant enough changes to multiple actions
+- Added color coding to indicate ranges in LOS overlay
+- Changed default to not toggling laying smoke after laying smoke
+- Bug fix: sponson units in range should get reaction fire
+- Bug fix: check that units can actually be towed when setting unit direction during deploy
 - Updated and promoted 106
 
 # 0.114